query testcase crud testcases update dependencies salesforceobject testcases added fixtures result testcases readme

+++ src/Client.php
@@ -12,7 +12,6 @@ use Closure,
   Throwable;
 
 use Nexcess\Salesforce\ {
-  Authentication\Authentication,
   Error\Salesforce as SalesforceException,
   Error\Usage as UsageException,
   Result,
@@ -29,7 +28,7 @@ use Psr\Http\Message\ {
 class Client {
 
   /** @var string Target Salesforce Api version. */
-  public const API_VERSION = 'v51.0';
+  public const API_VERSION = "v56.0";
 
   /**
    * Supported SOQL data types.
@@ -41,37 +40,34 @@ class Client {
    * @var string TYPE_BOOLEAN
    * @var string TYPE_NULL
    */
-  public const TYPE_STRING = 'string';
-  public const TYPE_INTEGER = 'integer';
-  public const TYPE_FLOAT = 'double';
-  public const TYPE_DECIMAL = 'decimal';
-  public const TYPE_BOOLEAN = 'boolean';
-  public const TYPE_NULL = 'NULL';
+  public const TYPE_STRING = "string";
+  public const TYPE_INTEGER = "integer";
+  public const TYPE_FLOAT = "double";
+  public const TYPE_DECIMAL = "decimal";
+  public const TYPE_BOOLEAN = "boolean";
+  public const TYPE_NULL = "NULL";
 
   /** @var string Base path for Api requests. */
-  protected const API_PATH = '/services/data/' . self::API_VERSION;
+  protected const API_PATH = "/services/data/" . self::API_VERSION;
 
   /**
    * @var string[] Literal value:escaped value map.
    *
-   * @see https://developer.salesforce.com/docs/atlas.en-us.soql_sosl.meta/
-   *  soql_sosl/sforce_api_calls_soql_select_quotedstringescapes.htm
+   * @see https://developer.salesforce.com/docs/atlas.en-us.soql_sosl.meta/soql_sosl/sforce_api_calls_soql_select_quotedstringescapes.htm
    */
   protected const ESCAPE_MAP = [
-    "\n" => '\n',
-    "\r" => '\r',
-    "\t" => '\t',
-    "\x07" => '\b',
-    "\f" => '\f',
-    '"' => '\"',
-    '\'' => '\\\'',
-    '\\' => '\\\\'
+    "\n" => "\\n",
+    "\r" => "\\r",
+    "\t" => "\\t",
+    "\x07" => "\b",
+    "\f" => "\\f",
+    "\"" => "\\\"",
+    "'" => "\\'",
+    "\\" => "\\\\"
   ];
 
   /**
    * Relevant HTTP status codes.
-   *
-   * @internal
    *
    * @var int HTTP_OK
    * @var int HTTP_CREATED
@@ -88,12 +84,12 @@ class Client {
   protected array $objectMap = [];
 
   /**
-   * @param Authentication $auth Authentication details
-   * @param string[] $object_map Map of salesforce object type:fully qualified classnames
+   * @param HttpClient $httpClient The authenticated Http Client to use
+   * @param string[] $objectMap Map of salesforce object type:fully qualified classnames
    */
-  public function __construct(Authentication $auth, array $object_map = []) {
-    $this->httpClient = $auth->httpClient();
-    $this->mapObjects($object_map);
+  public function __construct(HttpClient $httpClient, array $objectMap = []) {
+    $this->httpClient = $httpClient;
+    $this->mapObjects($objectMap);
   }
 
   /**
@@ -103,7 +99,7 @@ class Client {
    * @throws SalesforceException CREATE_FAILED on failure
    */
   public function create(SalesforceObject $object) : SalesforceObject {
-    $response = $this->request('POST', "/sobjects/{$object->type()}", ['json' => $object->toArray()]);
+    $response = $this->request("POST", "/sobjects/{$object->type()}", ["json" => $object->toArray()]);
     if ($response->getStatusCode() !== self::HTTP_CREATED) {
       throw SalesforceException::create(
         SalesforceException::CREATE_FAILED,
@@ -119,11 +115,19 @@ class Client {
   /**
    * Deletes an existing record in Salesforce.
    *
-   * @param SalesforceObject $object The object to create
+   * @param SalesforceObject $object The object to delete
+   * @throws UsageException EMPTY_ID If the $object cannot be deleted
    * @throws SalesforceException DELETE_FAILED on failure
    */
   public function delete(SalesforceObject $object) : SalesforceObject {
-    $this->deleteByExternalId($object->type(), 'Id', $object->Id);
+    if (empty($object->Id)) {
+      throw UsageException::create(
+        UsageException::EMPTY_ID,
+        ["type" => $object->type(), "id_field" => "Id", "object" => $object]
+      );
+    }
+
+    $this->deleteByExternalId($object->type(), "Id", $object->Id);
 
     $deleted = clone $object;
     $deleted->Id = null;
@@ -134,12 +138,12 @@ class Client {
    * Deletes an existing record in Salesforce given its type and an External Id.
    *
    * @param string $type Salesforce object type
-   * @param string $id_field The External Id field to look up by
+   * @param string $idField The External Id field to look up by
    * @param string $id 18-character Salesforce Id
    * @throws SalesforceException DELETE_FAILED on failure
    */
-  public function deleteByExternalId(string $type, string $id_field, string $id) : Result {
-    $response = $this->request('DELETE', "/sobjects/{$type}/{$id_field}/{$id}");
+  public function deleteByExternalId(string $type, string $idField, string $id) : Result {
+    $response = $this->request("DELETE", "/sobjects/{$type}/{$idField}/{$id}");
     if ($response->getStatusCode() !== self::HTTP_NO_CONTENT) {
       throw SalesforceException::create(
         SalesforceException::DELETE_FAILED,
@@ -158,19 +162,19 @@ class Client {
    * @throws SalesforceException GET_FAILED on failure
    */
   public function get(string $type, string $id) : SalesforceObject {
-    return $this->getByExternalId($type, 'Id', $id)->first();
+    return $this->getByExternalId($type, "Id", $id)->first();
   }
 
   /**
    * Gets a Salesforce record given its type and an External Id.
    *
    * @param string $type Salesforce object type
-   * @param string $id_field The External Id field to look up by
+   * @param string $idField The External Id field to look up by
    * @param string $id 18-character Salesforce Id
    * @throws SalesforceException GET_FAILED on failure
    */
-  public function getByExternalId(string $type, string $id_field, string $id) : Result {
-    $response = $this->request('GET', "/sobjects/{$type}/{$id_field}/{$id}");
+  public function getByExternalId(string $type, string $idField, string $id) : Result {
+    $response = $this->request("GET", "/sobjects/{$type}/{$idField}/{$id}");
     if ($response->getStatusCode() !== self::HTTP_OK) {
       throw SalesforceException::create(
         SalesforceException::GET_FAILED,
@@ -182,14 +186,28 @@ class Client {
   }
 
   /**
+   * Gets a new Result object for an Api Response.
+   *
+   * @param Response $response The Salesforce Api Response
+   * @return Result New Result object on success
+   */
+  public function getResultFrom(Response $response) : Result {
+    return Result::fromResponse(
+      $response,
+      $this->objectMap,
+      Closure::fromCallable([$this, "getMoreResultsFrom"])
+    );
+  }
+
+  /**
    * Maps Salesforce object types to desired php classnames.
    *
-   * @param string[] $object_map Map of salesforce object type:fully qualified classnames
+   * @param string[] $objectMap Map of salesforce object type:fully qualified classnames
    */
-  public function mapObjects(array $object_map) : void {
-    foreach ($object_map as $type => $fqcn) {
+  public function mapObjects(array $objectMap) : void {
+    foreach ($objectMap as $type => $fqcn) {
       if (! is_a($fqcn, SalesforceObject::class, true)) {
-        throw UsageException::create(UsageException::BAD_SFO_CLASSNAME, ['fqcn' => $fqcn]);
+        throw UsageException::create(UsageException::BAD_SFO_CLASSNAME, ["fqcn" => $fqcn]);
       }
 
       $this->objectMap[$type] = $fqcn;
@@ -201,14 +219,16 @@ class Client {
    *
    * @param string $template SOQL query with optional {tokens} for parameters
    * @param array $parameters Parameter values (escaped prior to formatting)
-   * @throws UsageException On error
+   * @throws UsageException If preparing or sending the request fails
+   * @throws SalesforceException If Salesforce response status is not 200 OK
    */
   public function query(string $template, array $parameters = []) : Result {
     $response = $this->request(
-      'GET',
-      '/query',
-      ['query' => ['q' => $this->prepare($template, $parameters)]]
+      "GET",
+      "/query",
+      ["query" => ["q" => $this->prepare($template, $parameters)]]
     );
+
     if ($response->getStatusCode() !== self::HTTP_OK) {
       throw SalesforceException::create(
         SalesforceException::GET_FAILED,
@@ -230,7 +250,7 @@ class Client {
     } catch (Throwable $e) {
       throw SalesforceException::create(
         SalesforceException::HTTP_REQUEST_FAILED,
-        ['method' => 'GET', 'path' => $path],
+        ["method" => "GET", "path" => $path],
         $e
       );
     }
@@ -241,34 +261,34 @@ class Client {
    *
    * @param SalesforceObject $object The object to update from
    * @throws SalesforceException UPDATE_FAILED on failure
-   * @throws UsageException NO_SUCH_FIELD If the object does not contain the given id_field
+   * @throws UsageException NO_SUCH_FIELD If the object does not contain the given idField
    */
   public function update(
     SalesforceObject $object,
-    string $id_field = "Id",
+    string $idField = "Id",
     string $id = null
   ) : SalesforceObject {
-    if (! property_exists($object, $id_field)) {
+    if (! property_exists($object, $idField)) {
       throw UsageException::create(
         UsageException::NO_SUCH_FIELD,
-        ['type' => $object->type(), 'field' => $id_field]
+        ["type" => $object->type(), "field" => $idField]
       );
     }
 
-    $id ??= $object->$id_field;
+    $id ??= $object->$idField;
     if (empty($id)) {
       throw UsageException::create(
         UsageException::EMPTY_ID,
-        ['type' => $object->type(), 'id_field' => $id_field]
+        ["type" => $object->type(), "idField" => $idField]
       );
     }
 
     $fields = $object->toArray();
-    unset($fields['Id']);
+    unset($fields["Id"]);
     $response = $this->request(
-      'PATCH',
-      "/sobjects/{$object->type()}/{$id_field}/{$id}",
-      ['json' => $fields]
+      "PATCH",
+      "/sobjects/{$object->type()}/{$idField}/{$id}",
+      ["json" => $fields]
     );
 
     if ($response->getStatusCode() !== self::HTTP_OK) {
@@ -292,38 +312,24 @@ class Client {
   }
 
   /**
-   * Gets a new Result object for an Api Response.
-   *
-   * @param Response $response The Salesforce Api Response
-   * @return Result New Result object on success
-   */
-  protected function getResultFrom(Response $response) : Result {
-    return Result::from(
-      $response,
-      $this->objectMap,
-      Closure::fromCallable([$this, 'getMoreResultsFrom'])
-    );
-  }
-
-  /**
    * Gets error context from an Api response.
    *
    * @param Response $response Http response to get context from
    */
   protected function getResponseContext(Response $response) : array {
-    $payload = json_decode($response->getBody()->__toString() ?: '[]')[0] ?? [];
+    $payload = json_decode($response->getBody()->__toString() ?: "[]")[0] ?? [];
 
     return array_filter(
       [
-        'response' => $response,
-        'status' => $response->getStatusCode(),
-        'reason' => $response->getReasonPhrase(),
-        'sf_id' => $payload->id ?? null,
-        'sf_errors' => $payload->errors ?? null,
-        'sf_success' => $payload->success ?? null,
-        'sf_created' => $payload->created ?? null,
-        'sf_error_code' => $payload->errorCode ?? null,
-        'sf_error_message' => isset($payload->errors) ?
+        "response" => $response,
+        "status" => $response->getStatusCode(),
+        "reason" => $response->getReasonPhrase(),
+        "sf_id" => $payload->id ?? null,
+        "sf_errors" => $payload->errors ?? null,
+        "sf_success" => $payload->success ?? null,
+        "sf_created" => $payload->created ?? null,
+        "sf_error_code" => $payload->errorCode ?? null,
+        "sf_error_message" => isset($payload->errors) ?
           join("\n", $payload->errors) :
           ($payload->message ?? null)
       ],
@@ -361,23 +367,23 @@ class Client {
    * @param string $type The intended datatype
    */
   protected function quote($value, string $type = null) : string {
-    $actual_type = gettype($value);
-    switch ($type ?? $actual_type) {
+    $actualType = gettype($value);
+    switch ($type ?? $actualType) {
       case self::TYPE_STRING:
-        if (! in_array($actual_type, [self::TYPE_STRING, self::TYPE_INTEGER, self::TYPE_FLOAT])) {
+        if (! in_array($actualType, [self::TYPE_STRING, self::TYPE_INTEGER, self::TYPE_FLOAT])) {
           throw UsageException::create(
             UsageException::UNQUOTABLE_VALUE,
-            ['type' => $type, 'actual_type' => $actual_type, 'value' => $value]
+            ["type" => $type, "actualType" => $actualType, "value" => $value]
           );
         }
 
-        return '\'' . strtr($value, self::ESCAPE_MAP) . '\'';
+        return "\"" . strtr($value, self::ESCAPE_MAP) . "\"";
       case self::TYPE_INTEGER:
         $integer = filter_var($value, FILTER_VALIDATE_INT);
         if ($integer === false) {
           throw UsageException::create(
             UsageException::UNQUOTABLE_VALUE,
-            ['type' => $type, 'actual_type' => $actual_type, 'value' => $value]
+            ["type" => $type, "actualType" => $actualType, "value" => $value]
           );
         }
 
@@ -387,44 +393,44 @@ class Client {
         if ($float === false) {
           throw UsageException::create(
             UsageException::UNQUOTABLE_VALUE,
-            ['type' => $type, 'actual_type' => $actual_type, 'value' => $value]
+            ["type" => $type, "actualType" => $actualType, "value" => $value]
           );
         }
 
         return $float;
       case self::TYPE_DECIMAL:
         if (
-          $actual_type !== self::TYPE_INTEGER &&
-          ! ($actual_type === self::TYPE_FLOAT && filter_var($value, FILTER_VALIDATE_INT) !== false) &&
-          ! ($actual_type === self::TYPE_STRING && filter_var($value, FILTER_VALIDATE_FLOAT) !== false)
+          $actualType !== self::TYPE_INTEGER &&
+          ! ($actualType === self::TYPE_FLOAT && filter_var($value, FILTER_VALIDATE_INT) !== false) &&
+          ! ($actualType === self::TYPE_STRING && filter_var($value, FILTER_VALIDATE_FLOAT) !== false)
         ) {
           throw UsageException::create(
             UsageException::UNQUOTABLE_VALUE,
-            ['type' => $type, 'actual_type' => $actual_type, 'value' => $value]
+            ["type" => $type, "actualType" => $actualType, "value" => $value]
           );
         }
 
         return (string) $value;
       case self::TYPE_BOOLEAN:
-        if ($actual_type !== self::TYPE_BOOLEAN) {
+        if ($actualType !== self::TYPE_BOOLEAN) {
           throw UsageException::create(
             UsageException::UNQUOTABLE_VALUE,
-            ['type' => $type, 'actual_type' => $actual_type, 'value' => $value]
+            ["type" => $type, "actualType" => $actualType, "value" => $value]
           );
         }
 
         return json_encode($value);
       case self::TYPE_NULL:
-        if ($actual_type !== self::TYPE_NULL) {
+        if ($actualType !== self::TYPE_NULL) {
           throw UsageException::create(
             UsageException::UNQUOTABLE_VALUE,
-            ['type' => $type, 'actual_type' => $actual_type, 'value' => $value]
+            ["type" => $type, "actualType" => $actualType, "value" => $value]
           );
         }
 
         return json_encode($value);
       default:
-        throw UsageException::create(UsageException::UNSUPPORTED_DATATYPE, ['type' => $type]);
+        throw UsageException::create(UsageException::UNSUPPORTED_DATATYPE, ["type" => $type]);
     }
   }
 
@@ -442,7 +448,7 @@ class Client {
     } catch (Throwable $e) {
       throw SalesforceException::create(
         SalesforceException::HTTP_REQUEST_FAILED,
-        ['method' => $method, 'path' => $path, 'options' => $options],
+        ["method" => $method, "path" => $path, "options" => $options],
         $e
       );
     }
